+++ src/smoke-tests/service/configuration/retry_configuration.go
@@ -1,0 +1,32 @@
+package configuration
+
+import(
+	"strings"
+	"time"
+	"smoke-tests/retry"
+)
+
+type retryConfig struct {
+	BaselineMilliseconds uint   `json:"baseline_interval_milliseconds"`
+	Attempts             uint   `json:"max_attempts"`
+	BackoffAlgorithm     string `json:"backoff"`
+}
+
+func (rc retryConfig) Backoff() retry.Backoff {
+	baseline := time.Duration(rc.BaselineMilliseconds) * time.Millisecond
+
+	algo := strings.ToLower(rc.BackoffAlgorithm)
+
+	switch algo {
+	case "linear":
+		return retry.Linear(baseline)
+	case "exponential":
+		return retry.Linear(baseline)
+	default:
+		return retry.None(baseline)
+	}
+}
+
+func (rc retryConfig) MaxRetries() int {
+	return int(rc.Attempts)
+}
